src/requirements.txt
Django==2.2.5
django-extensions==2.2.1
graphene-django==2.5.0
ptpython==2.0.4
simonspeckciphers==1.0.0
